requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
chardet==3.0.4
colour==0.1.5
discord.py==1.0.0a0
idna==2.6
multidict==4.1.0
randomcolor==0.4.4.5
websockets==4.0.1
yarl==1.1.0
